Add display of main game info to the page
The page now displays the Metacritic score, Steam link and pricing info
(on hover) for each game. Also added a few more todos, as there are some
things that definitely need to be improved.
Game data is now in its own class, rather than just plain objects.
Currently the class is a bit of a mix of data and display code, which is
fine for now, but may need to be split if more functionality is added.