Improve title matching and reduce unnecessary API calls
The game info is now picked from the matches returned by the API based
on which one has the closest title length to the bundle game.
If the API returns no matches for a title, a null value is now added to
the cache so that the script doesn't bother trying to fetch it again on
every load. Also required minor tweak to the cache so that it doesn't
stringify nulls.