Add /dist/ folder to repo for easier install & deployment
Should make it possible to use the webhook functions of GreasyFork and
OpenUserJS to update the script on those sites. Also gives somewhere for
the CSS file to be retrieved from (since the script hosting sites don't
seem to allow resouce hosting). Still need to actually set the CSS
location in the metadata though.