Rework things slightly so Steam deal links can replace a missing SteamID if needed
Fixes #2
Seems like this will be a pretty uncommon occurence, but one of the
games in the current weekly bundle has no SteamID, but does have a link
to a price on Steam. This commit means the Steam button will still be
shown is such situations, but will use the deal link rather than the
(missing) SteamID.

This also requires that current installs flush their cache in order for
the change to start working - this still needs to be implemented.